+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Features
+* Ported over the CLI & internal testsuite from `kurtosis-core`
 
 # 0.1.0
 * Initial commit